# openapi.yaml
openapi: 3.0.3
info:
  title: Secure API Gateway
  description: |
    ## Ключевые возможности
    * **Smart Load Balancing:** Round Robin распределение.
    * **Decentralized Health Checks:** Фоновые проверки бэкендов.
    * **JWT Security & Replay Protection:** Валидация JTI через Redis.
    * **Multi-Level Rate Limiting:** Лимиты в Redis.
    * **Dynamic IP Blacklist:** Блокировка через Redis.
    * **Payload Security:** Контроль размера Body.
    * **Logging:** Сквозное логирование с RequestID.
    * **JS Challenge:** Математическая задача для автоматической разблокировки IP.
  version: 1.0.0
  contact:
    name: veisdi, Weeelp

servers:
  - url: http://localhost:8080
    description: Локальный сервер разработки

tags:
  - name: Health
    description: Проверка состояния системы
  - name: Authenticated
    description: Требуют JWT токен
  - name: Challenge
    description: Защита от ботов (разблокировка IP)

paths:
  /health:
    get:
      tags: [Health]
      summary: Проверка здоровья сервера
      description: Возвращает OK, если хотя бы один бэкенд жив.
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                example: "OK"
        '503':
          description: Service Unavailable (все бэкенды мертвы)

  /:
    get:
      tags: [Authenticated]
      summary: Главная страница (прокси)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Успешный ответ от бэкенда
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Слишком много запросов

  /app:
    get:
      tags: [Authenticated]
      summary: Дополнительный эндпоинт (прокси)
      description: Не требует JWT, только проверку IP-блокировки.
      responses:
        '200':
          description: Ответ от бэкенда
        '403':
          $ref: '#/components/responses/Forbidden'

  /form:
    get:
      tags: [Authenticated]
      summary: Показать HTML форму
      security:
        - BearerAuth: []
      responses:
        '200':
          description: HTML форма
          content:
            text/html:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Authenticated]
      summary: Отправить данные формы
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Иван Иванов"
              required:
                - name
      responses:
        '200':
          description: Успешная обработка (прокси на бэкенд)
        '401':
          $ref: '#/components/responses/Unauthorized'

  /challenge:
    get:
      tags: [Challenge]
      summary: Генерация математической задачи
      description: |
        Возвращает HTML с простым арифметическим примером.
        Ответ сохраняется в Redis на 5 минут, привязан к IP пользователя.
      responses:
        '200':
          description: HTML страница с задачей
          content:
            text/html:
              schema:
                type: string
                example: "<h2>Решите задачу: 5 + 3</h2><form method='POST' action='/challenge/verify'><input name='answer' type='number'><button>Проверить</button></form>"

  /challenge/verify:
    post:
      tags: [Challenge]
      summary: Проверка ответа на задачу и разблокировка IP
      description: |
        При правильном ответе удаляет challenge и удаляет IP из чёрного списка.
        При неправильном – возвращает 403.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                answer:
                  type: string
                  example: "8"
              required:
                - answer
      responses:
        '200':
          description: Ответ верный, IP разблокирован
          content:
            text/html:
              schema:
                type: string
                example: "<h1>✅ Верно! Вы разблокированы.</h1>"
        '403':
          description: Неверный ответ, IP остаётся в чёрном списке
          content:
            text/html:
              schema:
                type: string
                example: "<h1>❌ Неверно. Попробуйте еще раз.</h1>"

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен для доступа к защищённым эндпоинтам

  responses:
    Unauthorized:
      description: Недопустимый или отсутствующий токен
      content:
        text/plain:
          schema:
            type: string
            example: "Invalid token"
    Forbidden:
      description: Доступ запрещён (IP в чёрном списке)
      content:
        text/plain:
          schema:
            type: string
            example: "Bot detected"